refactor(transformer): make previewer transformer replaceable
